requirements.txt
django-barcode
asgiref
astroid
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama
cryptography==3.3.1
defusedxml==0.6.0
Django==
django-allauth
django-countries
django-crispy-forms
django-debug-toolbar==3.2
django-filter
django-login-required-middleware
django-widget-tweaks
environ==1.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==2.0.0
pylint==2.4.4
python-decouple==3.4
python3-openid==3.2.0
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
stripe==2.55.1
urllib3==1.26.2
wrapt==1.11.2
